refactor: replace relative path to absolute path
