+++ src/atoms/modal.ts
@@ -1,5 +1,5 @@
 import { atom } from 'jotai';
 
-import { ModalProps } from '../components';
+import { ModalProps } from '@/components';
 
 export const modalsAtom = atom<ModalProps[]>([]);
