+++ src/atoms/toast.ts
@@ -1,5 +1,5 @@
 import { atom } from 'jotai';
 
-import { ToastProps } from '../components';
+import { ToastProps } from '@/components';
 
 export const toastsAtom = atom<ToastProps>({ visible: false, template: null });
