+++ src/components/index.ts
@@ -1,4 +1,6 @@
 export * from './common';
 export * from './layouts';
+export * from './modal';
+
 export * from './main';
 export * from './qna';
