+++ src/components/qna/QnASection/index.tsx
@@ -1,9 +1,5 @@
 import React, { useRef, useState } from 'react';
 
-import { AnimatePresence, LayoutGroup, motion } from 'framer-motion';
-
-import { QnAItem } from '@/constant/qnaMenu';
-
 import * as S from './styled';
 
 interface QnASectionProps {
