+++ src/hooks/index.ts
@@ -1,2 +1,4 @@
+export * from './useCounter';
 export * from './useModal';
+export * from './useScrollFadeIn';
 export * from './useToast';
