+++ src/pages/auth/signup/index.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useRef, useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { Link } from 'react-router-dom';
 
-import { SCHOOL } from '@/constant';
 import { Button, Input } from '@/components';
+import { SCHOOL } from '@/constant';
 import { useAuth } from '@/providers';
 
 import * as S from './styled';
